% tinyrox says don't edit this manually, but it can't stop you!
\name{apply_llama3_rope_scaling}
\alias{apply_llama3_rope_scaling}
\title{Apply Llama3-style RoPE scaling}
\usage{
apply_llama3_rope_scaling(inv_freq, scaling, dim)
}
\arguments{
\item{inv_freq}{Inverse frequencies}

\item{scaling}{Scaling configuration}

\item{dim}{Dimension}
}
\value{
Scaled inverse frequencies
}
\description{
Apply Llama3-style RoPE scaling
}
