% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{attention_block}
\alias{attention_block}
\title{Attention block for perceiver}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
attention_block
}
\description{
Attention block for perceiver
}
\keyword{datasets}
