% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/s3gen.R
\name{basic_transformer_block}
\alias{basic_transformer_block}
\title{Basic transformer block}
\usage{
basic_transformer_block(dim, num_heads = 8L)
}
\arguments{
\item{dim}{Hidden dimension}

\item{num_heads}{Number of attention heads}
}
\value{
nn_module
}
\description{
Basic transformer block
}
