% tinyrox says don't edit this manually, but it can't stop you!
\name{bfloat16_to_float32}
\alias{bfloat16_to_float32}
\title{Convert bfloat16 bytes to float32 values}
\usage{
bfloat16_to_float32(bytes)
}
\arguments{
\item{bytes}{Raw vector of bfloat16 bytes}
}
\value{
Numeric vector of float32 values
}
\description{
Convert bfloat16 bytes to float32 values
}
