% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{cam_dense_tdnn_layer}
\alias{cam_dense_tdnn_layer}
\title{CAM Dense TDNN Layer}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
cam_dense_tdnn_layer
}
\description{
CAM Dense TDNN Layer
}
\keyword{datasets}
