% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{cam_layer}
\alias{cam_layer}
\title{CAM (Context-Aware Masking) Layer}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
cam_layer
}
\description{
CAM (Context-Aware Masking) Layer
}
\keyword{datasets}
