% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/s3gen.R
\name{causal_conv1d}
\alias{causal_conv1d}
\title{Causal Conv1d - pads left only}
\usage{
causal_conv1d(
  in_channels,
  out_channels,
  kernel_size,
  stride = 1L,
  dilation = 1L
)
}
\arguments{
\item{in_channels}{Input channels}

\item{out_channels}{Output channels}

\item{kernel_size}{Kernel size}

\item{stride}{Stride (default 1)}

\item{dilation}{Dilation (default 1)}
}
\value{
nn_module
}
\description{
Causal Conv1d - pads left only
}
