% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{causal_resnet_block1d}
\alias{causal_resnet_block1d}
\title{Causal ResNet Block 1D}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
causal_resnet_block1d
}
\description{
Causal ResNet Block 1D
}
\keyword{datasets}
