% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{chat}
\alias{chat}
\title{Create a figure of a text conversation}
\usage{
chat(data, me, max_char_width = 80, font_size = 12, theme = "dark")
}
\arguments{
\item{data}{Expects a data.frame with columns "message_date", "text",
"sender_name"}

\item{me}{Who should be the person on the right? Should be a person in
data$sender_name}

\item{max_char_width}{How long the message be allowed to get (in characters)
before wrapping}

\item{font_size}{The font size, in points}

\item{theme}{Either a list (usually made by make_theme()) or a name of a
theme}
}
\value{
An image
}
\description{
Create a figure of a text conversation
}
\examples{

conversation |>
  chat(me = "Alice")
}
