% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{CHATTERBOX_REPO}
\alias{CHATTERBOX_REPO}
\title{Model Download Utilities}
\format{
An object of class \code{character} of length 1.
}
\usage{
CHATTERBOX_REPO
}
\description{
Download Chatterbox models from HuggingFace using hfhub.
Requires explicit download with user consent (no auto-download).
}
\keyword{datasets}
