% tinyrox says don't edit this manually, but it can't stop you!
\name{compute_mel_spectrogram}
\alias{compute_mel_spectrogram}
\title{Compute mel spectrogram (S3Gen compatible)}
\usage{
compute_mel_spectrogram(y, n_fft = 1920, n_mels = 80, sr = 24000,
                        hop_size = 480, win_size = 1920, fmin = 0, fmax = 8000,
                        center = FALSE)
}
\arguments{
\item{y}{Audio samples as torch tensor or numeric vector}

\item{n_fft}{FFT size (default 1920 for 24kHz)}

\item{n_mels}{Number of mel bins (default 80)}

\item{sr}{Sample rate (default 24000)}

\item{hop_size}{Hop size (default 480)}

\item{win_size}{Window size (default 1920)}

\item{fmin}{Minimum frequency (default 0)}

\item{fmax}{Maximum frequency (default 8000)}

\item{center}{Whether to center frames (default FALSE)}
}
\value{
Mel spectrogram tensor (batch, n_mels, time)
}
\description{
Compute mel spectrogram (S3Gen compatible)
}
\keyword{internal}
