% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voice_encoder.R
\name{compute_ve_mel}
\alias{compute_ve_mel}
\title{Compute mel spectrogram for voice encoder}
\usage{
compute_ve_mel(wav, config = voice_encoder_config())
}
\arguments{
\item{wav}{Audio samples (numeric vector)}

\item{config}{Voice encoder config}
}
\value{
Mel spectrogram (time, n_mels)
}
\description{
Compute mel spectrogram for voice encoder
}
