% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{conformer_encoder_layer}
\alias{conformer_encoder_layer}
\title{Conformer Encoder Layer}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
conformer_encoder_layer
}
\description{
Single conformer block with attention and feed-forward (no convolution).
}
\keyword{datasets}
