% tinyrox says don't edit this manually, but it can't stop you!
\name{create_s3gen_vocoder}
\alias{create_s3gen_vocoder}
\title{Create HiFiGAN vocoder with S3Gen configuration}
\usage{
create_s3gen_vocoder(device = "cpu")
}
\arguments{
\item{device}{Target device}
}
\value{
HiFTGenerator module
}
\description{
Create HiFiGAN vocoder with S3Gen configuration
}
\keyword{internal}
