% tinyrox says don't edit this manually, but it can't stop you!
\name{dense_layer}
\alias{dense_layer}
\title{Dense layer for final embedding}
\usage{
dense_layer(in_channels, out_channels)
}
\arguments{
\item{in_channels}{Input channels}

\item{out_channels}{Output channels}
}
\value{
nn_module
}
\description{
Dense layer for final embedding
}
