% tinyrox says don't edit this manually, but it can't stop you!
\name{.gpt2_bytes_to_unicode}
\alias{.gpt2_bytes_to_unicode}
\title{GPT-2 bytes-to-unicode mapping}
\usage{
.gpt2_bytes_to_unicode()
}
\value{
Named character vector (byte value as character -> unicode char)
}
\description{
GPT-2 bytes-to-unicode mapping
}
\keyword{internal}
