% tinyrox says don't edit this manually, but it can't stop you!
\name{fcm_module}
\alias{fcm_module}
\title{Factorized Convolutional Module (FCM)}
\usage{
fcm_module(m_channels = 32, feat_dim = 80)
}
\arguments{
\item{m_channels}{Number of channels}

\item{feat_dim}{Input feature dimension (mel bins)}
}
\value{
nn_module
}
\description{
Factorized Convolutional Module (FCM)
}
