% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/s3tokenizer.R
\name{fsmn_multi_head_attention}
\alias{fsmn_multi_head_attention}
\title{FSMN Multi-Head Attention}
\arguments{
\item{n_state}{Hidden dimension}

\item{n_head}{Number of heads}

\item{kernel_size}{FSMN kernel size (default 31)}
}
\value{
nn_module
}
\description{
Multi-head attention with Frequency-domain Self-attention Memory Network
}
