% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/s3tokenizer.R
\name{fsq_vector_quantization}
\alias{fsq_vector_quantization}
\title{FSQ Vector Quantization wrapper}
\usage{
fsq_vector_quantization(dim, codebook_size = 6561L)
}
\arguments{
\item{dim}{Input dimension}

\item{codebook_size}{Codebook size (must be 6561 = 3^8)}
}
\value{
nn_module
}
\description{
FSQ Vector Quantization wrapper
}
