% tinyrox says don't edit this manually, but it can't stop you!
\name{get_dtype_info}
\alias{get_dtype_info}
\title{Get dtype information}
\usage{
get_dtype_info(dtype)
}
\arguments{
\item{dtype}{Safetensors dtype string}
}
\value{
List with r_type, size, torch_dtype
}
\description{
Get dtype information
}
