% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{get_theme}
\alias{get_theme}
\title{Return a theme by its name}
\usage{
get_theme(theme = c("dark", "light"))
}
\arguments{
\item{theme}{Character}
}
\value{
A list, containing all necessary components of a theme
}
\description{
Return a theme by its name
}
