% tinyrox says don't edit this manually, but it can't stop you!
\name{get_turbo_model_paths}
\alias{get_turbo_model_paths}
\title{Get Paths to Downloaded Turbo Model Files}
\usage{
get_turbo_model_paths(revision = NULL)
}
\arguments{
\item{revision}{Optional exact 40-hex commit to resolve against.}
}
\value{
Named list of local file paths
}
\description{
Get Paths to Downloaded Turbo Model Files
}
\keyword{internal}
