% tinyrox says don't edit this manually, but it can't stop you!
\name{gpt2_attention}
\alias{gpt2_attention}
\title{GPT-2 Attention (combined QKV projection)}
\usage{
gpt2_attention(config)
}
\arguments{
\item{config}{GPT-2 config}
}
\value{
nn_module
}
\description{
GPT-2 Attention (combined QKV projection)
}
