% tinyrox says don't edit this manually, but it can't stop you!
\name{gpt2_layer_norm}
\alias{gpt2_layer_norm}
\title{GPT-2 Layer Normalization}
\usage{
gpt2_layer_norm(hidden_size, eps = 1e-05)
}
\arguments{
\item{hidden_size}{Dimension}

\item{eps}{Epsilon}
}
\value{
nn_module
}
\description{
GPT-2 Layer Normalization
}
