% tinyrox says don't edit this manually, but it can't stop you!
\name{gpt2_model}
\alias{gpt2_model}
\title{GPT-2 Model (transformer backbone)}
\usage{
gpt2_model(config = NULL)
}
\arguments{
\item{config}{GPT-2 configuration}
}
\value{
nn_module
}
\description{
GPT-2 Model (transformer backbone)
}
