% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{hifigan_resblock}
\alias{hifigan_resblock}
\title{HiFiGAN Residual Block}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
hifigan_resblock
}
\description{
HiFiGAN Residual Block
}
\keyword{datasets}
