% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{hift_generator}
\alias{hift_generator}
\title{HiFTNet Generator}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
hift_generator
}
\description{
Neural Source Filter + ISTFTNet
Reference: https://arxiv.org/abs/2309.09493
}
\keyword{datasets}
