% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/s3gen.R
\name{inference}
\alias{inference}
\title{Run inference (tokens -> mel -> audio)}
\usage{
inference(
  speech_tokens,
  ref_wav = NULL,
  ref_sr = NULL,
  ref_dict = NULL,
  finalize = TRUE
)
}
\description{
Run inference (tokens -> mel -> audio)
}
