% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/tts.R
\name{is_loaded}
\alias{is_loaded}
\title{Check if model is loaded}
\usage{
is_loaded(model)
}
\arguments{
\item{model}{Chatterbox model}
}
\value{
TRUE if model is loaded
}
\description{
Check if model is loaded
}
