% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/conformer.R
\name{linear_no_subsampling}
\alias{linear_no_subsampling}
\title{Linear No Subsampling layer}
\usage{
linear_no_subsampling(input_dim = 512, output_dim = 512, dropout_rate = 0.1)
}
\arguments{
\item{input_dim}{Input dimension}

\item{output_dim}{Output dimension}

\item{dropout_rate}{Dropout rate}
}
\value{
nn_module
}
\description{
Projects input to model dimension with layer norm and positional encoding.
}
