% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_model}
\alias{llama_model}
\title{Llama model (decoder only)}
\usage{
llama_model(config = NULL)
}
\arguments{
\item{config}{Model configuration (default: 520M)}
}
\value{
nn_module
}
\description{
Llama model (decoder only)
}
