% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{llama_rms_norm}
\alias{llama_rms_norm}
\title{RMS Normalization module}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
llama_rms_norm
}
\description{
RMS Normalization module
}
\keyword{datasets}
