% tinyrox says don't edit this manually, but it can't stop you!
\name{load_conformer_encoder_weights}
\alias{load_conformer_encoder_weights}
\title{Load Conformer Encoder weights}
\usage{
load_conformer_encoder_weights(model, state_dict, prefix = "flow.encoder.")
}
\arguments{
\item{model}{Conformer encoder module}

\item{state_dict}{State dictionary}

\item{prefix}{Key prefix (e.g., "flow.encoder.")}
}
\value{
The \code{model} module, with weights copied in from
  \code{state_dict}.
}
\description{
Load Conformer Encoder weights
}
