% tinyrox says don't edit this manually, but it can't stop you!
\name{load_hifigan_weights}
\alias{load_hifigan_weights}
\title{Load HiFiGAN weights from state dictionary}
\usage{
load_hifigan_weights(model, state_dict, prefix = "mel2wav.")
}
\arguments{
\item{model}{HiFTGenerator model}

\item{state_dict}{Named list of tensors}

\item{prefix}{Prefix for weight keys (default "mel2wav.")}
}
\value{
Model with loaded weights
}
\description{
Load HiFiGAN weights from state dictionary
}
\keyword{internal}
