% tinyrox says don't edit this manually, but it can't stop you!
\name{load_t3_weights}
\alias{load_t3_weights}
\title{Load T3 weights from safetensors}
\usage{
load_t3_weights(model, state_dict)
}
\arguments{
\item{model}{T3 model}

\item{state_dict}{Named list of tensors}
}
\value{
Model with loaded weights
}
\description{
Load T3 weights from safetensors
}
