% tinyrox says don't edit this manually, but it can't stop you!
\name{load_tokenizer}
\alias{load_tokenizer}
\title{Load tokenizer from JSON file (internal)}
\usage{
load_tokenizer(vocab_path)
}
\arguments{
\item{vocab_path}{Path to tokenizer.json}
}
\value{
Tokenizer object (list)
}
\description{
Load tokenizer from JSON file (internal)
}
