% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{make_theme}
\alias{make_theme}
\title{Make a custom message theme}
\usage{
make_theme(
  text_me = "white",
  text_you = "black",
  fill_me = "#218aff",
  fill_you = "gray80",
  bg = "white"
)
}
\arguments{
\item{text_me}{The color of text of the person on the right}

\item{text_you}{The color of the text of the person on the left}

\item{fill_me}{The color of the bubble of the person on the right}

\item{fill_you}{The color of the bubble of the person on the left}

\item{bg}{The color of the background}
}
\value{
A list
}
\description{
Make a custom message theme
}
