% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msgGrob}
\alias{msgGrob}
\title{Create a single msg graphical object}
\usage{
msgGrob(
  text,
  is_me = TRUE,
  max_char_width = 80,
  font_size = 12,
  theme = "light"
)
}
\arguments{
\item{text}{The message to be displayed}

\item{is_me}{Should this look like the sender or the receiver?}

\item{max_char_width}{Number of characters until wrapping}

\item{font_size}{The font size, in points}

\item{theme}{Either a character representing one of the themes in
"get_theme", or a list (usually created from make_theme)}
}
\value{
A grid::gTree
}
\description{
Create a single msg graphical object
}
\examples{
msgGrob("Hello!") |>
  grid::grid.draw()
}
