% tinyrox says don't edit this manually, but it can't stop you!
\name{pad_audio_for_tokenizer}
\alias{pad_audio_for_tokenizer}
\title{Pad audio to multiple of token rate}
\usage{
pad_audio_for_tokenizer(wav, sr)
}
\arguments{
\item{wav}{Audio samples}

\item{sr}{Sample rate}
}
\value{
Padded audio
}
\description{
Pad audio to multiple of token rate
}
