% tinyrox says don't edit this manually, but it can't stop you!
\name{print.chatterbox}
\alias{print.chatterbox}
\title{Print method for chatterbox}
\usage{
\method{print}{chatterbox}(x, ...)
}
\arguments{
\item{x}{Chatterbox model}

\item{...}{Ignored}
}
\value{
\code{x}, invisibly. Called for the side effect of printing a
  summary of the model to the console.
}
\description{
Print method for chatterbox
}
