% tinyrox says don't edit this manually, but it can't stop you!
\name{print.chatterbox_defaults}
\alias{print.chatterbox_defaults}
\title{Print method for chatterbox_defaults}
\usage{
\method{print}{chatterbox_defaults}(x, ...)
}
\arguments{
\item{x}{Object from \code{\link{chatterbox_defaults}}}

\item{...}{Ignored}
}
\value{
\code{x}, invisibly
}
\description{
Print method for chatterbox_defaults
}
