% tinyrox says don't edit this manually, but it can't stop you!
\name{print.chatterbox_gc_options}
\alias{print.chatterbox_gc_options}
\title{Print method for chatterbox_gc_options}
\usage{
\method{print}{chatterbox_gc_options}(x, ...)
}
\arguments{
\item{x}{Object from \code{\link{chatterbox_gc_options}}}

\item{...}{Ignored}
}
\value{
\code{x}, invisibly
}
\description{
Print method for chatterbox_gc_options
}
