% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{punc_norm}
\alias{punc_norm}
\title{Normalize punctuation for TTS}
\usage{
punc_norm(text)
}
\arguments{
\item{text}{Input text}
}
\value{
Normalized text
}
\description{
Normalize punctuation for TTS
}
