% tinyrox says don't edit this manually, but it can't stop you!
\name{reflection_pad1d}
\alias{reflection_pad1d}
\title{Reflection padding for 1D (nn_reflection_pad1d equivalent)}
\usage{
reflection_pad1d(padding)
}
\arguments{
\item{padding}{Integer vector c(left, right) for padding}
}
\value{
nn_module
}
\description{
Reflection padding for 1D (nn_reflection_pad1d equivalent)
}
