% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{rotate_half}
\alias{rotate_half}
\title{Rotate half of the tensor for RoPE}
\usage{
rotate_half(x)
}
\arguments{
\item{x}{Input tensor}
}
\value{
Rotated tensor
}
\description{
Rotate half of the tensor for RoPE
}
