% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/s3tokenizer.R
\name{s3_audio_encoder}
\alias{s3_audio_encoder}
\title{S3 Audio Encoder V2}
\arguments{
\item{n_mels}{Number of mel bins}

\item{n_state}{Hidden dimension}

\item{n_head}{Number of attention heads}

\item{n_layer}{Number of transformer layers}

\item{stride}{Convolution stride (default 2)}
}
\value{
nn_module
}
\description{
S3 Audio Encoder V2
}
