% tinyrox says don't edit this manually, but it can't stop you!
\name{s3_multi_head_attention}
\alias{s3_multi_head_attention}
\title{Multi-Head Attention base module}
\usage{
s3_multi_head_attention(n_state, n_head)
}
\arguments{
\item{n_state}{Hidden dimension}

\item{n_head}{Number of heads}
}
\value{
nn_module
}
\description{
Multi-Head Attention base module
}
