% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/s3tokenizer.R
\name{s3_tokenizer_config}
\alias{s3_tokenizer_config}
\title{S3Tokenizer model configuration}
\usage{
s3_tokenizer_config(
  n_mels = 128,
  n_audio_state = 1280,
  n_audio_head = 20,
  n_audio_layer = 6,
  n_codebook_size = 6561
)
}
\arguments{
\item{n_mels}{Number of mel bins (default 128)}

\item{n_audio_state}{Hidden state dimension (default 1280)}

\item{n_audio_head}{Number of attention heads (default 20)}

\item{n_audio_layer}{Number of transformer layers (default 6)}

\item{n_codebook_size}{Codebook size (default 6561 = 3^8)}
}
\value{
Configuration list
}
\description{
S3Tokenizer model configuration
}
