% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{sine_gen}
\alias{sine_gen}
\title{Sine Generator}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
sine_gen
}
\description{
Generates sine waveforms from F0 for source-filter synthesis
}
\keyword{datasets}
