% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/s3gen.R
\name{sinusoidal_pos_emb}
\alias{sinusoidal_pos_emb}
\title{Sinusoidal positional embedding for timesteps}
\usage{
sinusoidal_pos_emb(dim = 320L)
}
\arguments{
\item{dim}{Output dimension}
}
\value{
nn_module
}
\description{
Sinusoidal positional embedding for timesteps
}
