% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{snake_activation}
\alias{snake_activation}
\title{Snake activation function}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
snake_activation
}
\description{
Sine-based periodic activation: x + 1/a * sin^2(ax)
Reference: https://arxiv.org/abs/2006.08195
}
\keyword{datasets}
