% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{source_module_hn_nsf}
\alias{source_module_hn_nsf}
\title{Source Module for Neural Source Filter}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
source_module_hn_nsf
}
\description{
Source Module for Neural Source Filter
}
\keyword{datasets}
