% tinyrox says don't edit this manually, but it can't stop you!
\name{statistics_pooling}
\alias{statistics_pooling}
\title{Statistics pooling}
\usage{
statistics_pooling(x)
}
\arguments{
\item{x}{Input tensor (batch, channels, time)}
}
\value{
Statistics tensor (batch, channels * 2)
}
\description{
Statistics pooling
}
