% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{t3_cond_enc}
\alias{t3_cond_enc}
\title{T3 conditioning encoder}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
t3_cond_enc
}
\description{
T3 conditioning encoder
}
\keyword{datasets}
