% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/t3.R
\name{t3_cond_to_device}
\alias{t3_cond_to_device}
\title{Move T3 conditioning to device}
\usage{
t3_cond_to_device(cond, device)
}
\arguments{
\item{cond}{T3 conditioning object}

\item{device}{Target device}
}
\value{
T3 conditioning on device
}
\description{
Move T3 conditioning to device
}
