% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/t3.R
\name{t3_config_english}
\alias{t3_config_english}
\title{Create T3 configuration (English-only)}
\usage{
t3_config_english()
}
\value{
List with T3 configuration
}
\description{
Create T3 configuration (English-only)
}
