% tinyrox says don't edit this manually, but it can't stop you!
\name{t3_config_turbo}
\alias{t3_config_turbo}
\title{Create T3 turbo configuration (GPT-2 backbone)}
\usage{
t3_config_turbo()
}
\value{
List with T3 turbo configuration
}
\description{
Create T3 turbo configuration (GPT-2 backbone)
}
