% tinyrox says don't edit this manually, but it can't stop you!
\name{t3_inference}
\alias{t3_inference}
\title{Run T3 inference to generate speech tokens}
\usage{
t3_inference(
  model,
  cond,
  text_tokens,
  max_new_tokens = 1000,
  temperature = 0.8,
  cfg_weight = 0.5,
  top_p = 1,
  min_p = 0.05,
  repetition_penalty = 1.2
)
}
\arguments{
\item{model}{T3 model}

\item{cond}{T3 conditioning}

\item{text_tokens}{Tokenized text (tensor)}

\item{max_new_tokens}{Maximum speech tokens to generate}

\item{temperature}{Sampling temperature}

\item{cfg_weight}{Classifier-free guidance weight}

\item{top_p}{Nucleus sampling threshold (1.0 disables, Python default)}

\item{min_p}{Minimum probability threshold}

\item{repetition_penalty}{Repetition penalty}
}
\value{
Generated speech tokens
}
\description{
Run T3 inference to generate speech tokens
}
\keyword{internal}
