% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{t3_model}
\alias{t3_model}
\title{T3 Token-to-Token TTS model}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
t3_model
}
\description{
T3 Token-to-Token TTS model
}
\keyword{datasets}
