% tinyrox says don't edit this manually, but it can't stop you!
\name{t3_model_turbo}
\alias{t3_model_turbo}
\title{T3 Token-to-Token TTS model (Turbo variant with GPT-2 backbone)}
\usage{
t3_model_turbo(config = NULL)
}
\arguments{
\item{config}{T3 turbo configuration}
}
\value{
nn_module
}
\description{
T3 Token-to-Token TTS model (Turbo variant with GPT-2 backbone)
}
