% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/speaker_encoder.R
\name{tdnn_layer}
\alias{tdnn_layer}
\title{TDNN Layer}
\arguments{
\item{in_channels}{Input channels}

\item{out_channels}{Output channels}

\item{kernel_size}{Kernel size}

\item{stride}{Stride}

\item{dilation}{Dilation}

\item{padding}{Padding (default: computed from kernel_size and dilation)}
}
\value{
nn_module
}
\description{
TDNN Layer
}
