% tinyrox says don't edit this manually, but it can't stop you!
\name{text_to_tokens}
\alias{text_to_tokens}
\title{Convert text to token tensor}
\usage{
text_to_tokens(tokenizer, text, normalize = TRUE, device = "cpu")
}
\arguments{
\item{tokenizer}{Tokenizer object}

\item{text}{Input text}

\item{normalize}{Whether to apply punctuation normalization}

\item{device}{Target device}
}
\value{
Token tensor (1, seq_len)
}
\description{
Convert text to token tensor
}
\keyword{internal}
