% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{timestep_embedding}
\alias{timestep_embedding}
\title{Timestep embedding MLP}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
timestep_embedding
}
\description{
Timestep embedding MLP
}
\keyword{datasets}
