% tinyrox says don't edit this manually, but it can't stop you!
\name{tokenize_text_gpt2}
\alias{tokenize_text_gpt2}
\title{Tokenize text using GPT-2 BPE}
\usage{
tokenize_text_gpt2(tokenizer, text)
}
\arguments{
\item{tokenizer}{GPT-2 tokenizer from load_gpt2_tokenizer()}

\item{text}{Input text}
}
\value{
Integer vector of token IDs
}
\description{
Tokenize text using GPT-2 BPE
}
\keyword{internal}
