% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{traceable_transformer_first}
\alias{traceable_transformer_first}
\title{Traceable transformer for first token (no cache)}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
traceable_transformer_first
}
\description{
Traceable transformer for first token (no cache)
}
\keyword{datasets}
