% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{transit_layer}
\alias{transit_layer}
\title{Transit layer (channel reduction)}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
transit_layer
}
\description{
Transit layer (channel reduction)
}
\keyword{datasets}
