% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{transpose_layer}
\alias{transpose_layer}
\title{Transpose layer for use in sequential}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
transpose_layer
}
\description{
Transpose layer for use in sequential
}
\keyword{datasets}
