% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{upsample_conformer_encoder_full}
\alias{upsample_conformer_encoder_full}
\title{Upsample Conformer Encoder}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
upsample_conformer_encoder_full
}
\description{
Full conformer encoder matching Python UpsampleConformerEncoder.
}
\keyword{datasets}
