% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{voice_encoder}
\alias{voice_encoder}
\title{Voice encoder module}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
voice_encoder
}
\description{
Voice encoder module
}
\keyword{datasets}
