% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/voice_encoder.R
\name{voice_encoder_config}
\alias{voice_encoder_config}
\title{Voice encoder configuration}
\usage{
voice_encoder_config()
}
\value{
List with configuration parameters
}
\description{
Voice encoder configuration
}
