% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/t3.R
\name{attention_block}
\alias{attention_block}
\title{Attention block for perceiver}
\usage{
attention_block(embed_dim = 1024, num_heads = 4)
}
\arguments{
\item{embed_dim}{Embedding dimension (default 1024)}

\item{num_heads}{Number of attention heads (default 4)}
}
\value{
nn_module
}
\description{
Attention block for perceiver
}
