% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{basic_res_block}
\alias{basic_res_block}
\title{Basic residual block for FCM}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
basic_res_block
}
\description{
Basic residual block for FCM
}
\keyword{datasets}
