% tinyrox says don't edit this manually, but it can't stop you!
\name{basic_res_block}
\alias{basic_res_block}
\title{Basic residual block for FCM}
\usage{
basic_res_block(in_planes, planes, stride = 1)
}
\arguments{
\item{in_planes}{Input channels}

\item{planes}{Output channels}

\item{stride}{Stride for downsampling}
}
\value{
nn_module
}
\description{
Basic residual block for FCM
}
