% tinyrox says don't edit this manually, but it can't stop you!
\name{basic_transformer_block}
\alias{basic_transformer_block}
\title{Basic transformer block}
\usage{
basic_transformer_block(dim, num_heads = 8L)
}
\arguments{
\item{dim}{Hidden dimension}

\item{num_heads}{Number of attention heads}
}
\value{
nn_module
}
\description{
Basic transformer block
}
