% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{basic_transformer_block}
\alias{basic_transformer_block}
\title{Basic transformer block}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
basic_transformer_block
}
\description{
Basic transformer block
}
\keyword{datasets}
