% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/speaker_encoder.R
\name{cam_dense_tdnn_block}
\alias{cam_dense_tdnn_block}
\title{CAM Dense TDNN Block (multiple layers with dense connections)}
\usage{
cam_dense_tdnn_block(
  num_layers,
  in_channels,
  out_channels,
  bn_channels,
  kernel_size,
  dilation = 1
)
}
\arguments{
\item{num_layers}{Number of layers}

\item{in_channels}{Input channels}

\item{out_channels}{Output channels per layer}

\item{bn_channels}{Bottleneck channels}

\item{kernel_size}{Kernel size}

\item{dilation}{Dilation}
}
\value{
nn_module
}
\description{
CAM Dense TDNN Block (multiple layers with dense connections)
}
