% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{cam_dense_tdnn_block}
\alias{cam_dense_tdnn_block}
\title{CAM Dense TDNN Block (multiple layers with dense connections)}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
cam_dense_tdnn_block
}
\description{
CAM Dense TDNN Block (multiple layers with dense connections)
}
\keyword{datasets}
