% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/speaker_encoder.R
\name{cam_dense_tdnn_layer}
\alias{cam_dense_tdnn_layer}
\title{CAM Dense TDNN Layer}
\usage{
cam_dense_tdnn_layer(
  in_channels,
  out_channels,
  bn_channels,
  kernel_size,
  dilation = 1
)
}
\arguments{
\item{in_channels}{Input channels}

\item{out_channels}{Output channels}

\item{bn_channels}{Bottleneck channels}

\item{kernel_size}{Kernel size}

\item{dilation}{Dilation}
}
\value{
nn_module
}
\description{
CAM Dense TDNN Layer
}
