% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/speaker_encoder.R
\name{cam_layer}
\alias{cam_layer}
\title{CAM (Context-Aware Masking) Layer}
\arguments{
\item{bn_channels}{Bottleneck channels}

\item{out_channels}{Output channels}

\item{kernel_size}{Kernel size}

\item{stride}{Stride}

\item{padding}{Padding}

\item{dilation}{Dilation}

\item{reduction}{Channel reduction factor}
}
\value{
nn_module
}
\description{
CAM (Context-Aware Masking) Layer
}
