% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{causal_block1d}
\alias{causal_block1d}
\title{Causal Block 1D - CausalConv + LayerNorm + Mish}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
causal_block1d
}
\description{
Causal Block 1D - CausalConv + LayerNorm + Mish
}
\keyword{datasets}
