% tinyrox says don't edit this manually, but it can't stop you!
\name{causal_block1d}
\alias{causal_block1d}
\title{Causal Block 1D - CausalConv + LayerNorm + Mish}
\usage{
causal_block1d(in_channels, out_channels, kernel_size = 3L)
}
\arguments{
\item{in_channels}{Input channels}

\item{out_channels}{Output channels}

\item{kernel_size}{Kernel size}
}
\value{
nn_module
}
\description{
Causal Block 1D - CausalConv + LayerNorm + Mish
}
