% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3gen.R
\name{causal_cfm}
\alias{causal_cfm}
\title{Causal Conditional Flow Matching}
\usage{
causal_cfm(in_channels = 320, out_channels = 80, spk_emb_dim = 80)
}
\arguments{
\item{in_channels}{Input channels (x + mu + spks + cond)}

\item{out_channels}{Output channels (mel bins)}

\item{spk_emb_dim}{Speaker embedding dimension}
}
\value{
nn_module
}
\description{
Causal Conditional Flow Matching
}
