% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{causal_conv1d}
\alias{causal_conv1d}
\title{Causal Conv1d - pads left only}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
causal_conv1d
}
\description{
Causal Conv1d - pads left only
}
\keyword{datasets}
