% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/s3gen.R
\name{causal_masked_diff_xvec}
\alias{causal_masked_diff_xvec}
\title{Causal Masked Diff with Xvector}
\arguments{
\item{vocab_size}{Speech token vocabulary size}

\item{input_size}{Token embedding size}

\item{output_size}{Mel bins}

\item{spk_embed_dim}{Speaker embedding dimension}

\item{input_frame_rate}{Input frame rate for audio processing}

\item{token_mel_ratio}{Ratio of tokens to mel frames}
}
\value{
nn_module
}
\description{
Causal Masked Diff with Xvector
}
