% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{causal_masked_diff_xvec}
\alias{causal_masked_diff_xvec}
\title{Causal Masked Diff with Xvector}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
causal_masked_diff_xvec
}
\description{
Causal Masked Diff with Xvector
}
\keyword{datasets}
