% tinyrox says don't edit this manually, but it can't stop you!
\name{causal_resnet_block1d}
\alias{causal_resnet_block1d}
\title{Causal ResNet Block 1D}
\usage{
causal_resnet_block1d(in_channels, out_channels, time_embed_dim = 1024L)
}
\arguments{
\item{in_channels}{Input channels}

\item{out_channels}{Output channels}

\item{time_embed_dim}{Time embedding dimension}
}
\value{
nn_module
}
\description{
Causal ResNet Block 1D
}
