% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{cfm_attention}
\alias{cfm_attention}
\title{Self-attention for transformer block}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
cfm_attention
}
\description{
Self-attention for transformer block
}
\keyword{datasets}
