% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/s3gen.R
\name{cfm_estimator}
\alias{cfm_estimator}
\title{CFM Estimator (UNet1D style)}
\usage{
cfm_estimator(in_channels = 320, out_channels = 80)
}
\arguments{
\item{in_channels}{Input channels}

\item{out_channels}{Output channels}
}
\value{
nn_module
}
\description{
CFM Estimator (UNet1D style)
}
