% tinyrox says don't edit this manually, but it can't stop you!
\name{chatterbox}
\alias{chatterbox}
\title{Create Chatterbox TTS model}
\usage{
chatterbox(device = "cpu")
}
\arguments{
\item{device}{Device to use ("cpu", "cuda", "mps", etc.)}
}
\value{
Chatterbox TTS model object
}
\description{
Create Chatterbox TTS model
}
