% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.R
\name{chatterbox}
\alias{chatterbox}
\title{Create Chatterbox TTS model}
\usage{
chatterbox(device = "cpu")
}
\arguments{
\item{device}{Device to use ("cpu", "cuda", "mps", etc.)}
}
\value{
Chatterbox TTS model object
}
\description{
Create Chatterbox TTS model
}
