% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/hifigan.R
\name{conv_rnn_f0_predictor}
\alias{conv_rnn_f0_predictor}
\title{Convolutional RNN F0 Predictor}
\usage{
conv_rnn_f0_predictor(in_channels = 80, cond_channels = 512)
}
\arguments{
\item{in_channels}{Input channels (mel bins)}

\item{cond_channels}{Hidden channels}
}
\value{
nn_module
}
\description{
Convolutional RNN F0 Predictor
}
