% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{conv_rnn_f0_predictor}
\alias{conv_rnn_f0_predictor}
\title{Convolutional RNN F0 Predictor}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
conv_rnn_f0_predictor
}
\description{
Convolutional RNN F0 Predictor
}
\keyword{datasets}
