% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hifigan.R
\name{create_s3gen_vocoder}
\alias{create_s3gen_vocoder}
\title{Create HiFiGAN vocoder with S3Gen configuration}
\usage{
create_s3gen_vocoder(device = "cpu")
}
\arguments{
\item{device}{Target device}
}
\value{
HiFTGenerator module
}
\description{
Create HiFiGAN vocoder with S3Gen configuration
}
