% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{decode_tokens}
\alias{decode_tokens}
\title{Decode token IDs to text}
\usage{
decode_tokens(tokenizer, ids)
}
\arguments{
\item{tokenizer}{Tokenizer object}

\item{ids}{Integer vector or tensor of token IDs}
}
\value{
Decoded text string
}
\description{
Decode token IDs to text
}
