% tinyrox says don't edit this manually, but it can't stop you!
\name{decode_tokens}
\alias{decode_tokens}
\title{Decode token IDs to text}
\usage{
decode_tokens(tokenizer, ids)
}
\arguments{
\item{tokenizer}{Tokenizer object}

\item{ids}{Integer vector or tensor of token IDs}
}
\value{
Decoded text string
}
\description{
Decode token IDs to text
}
