% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{dense_layer}
\alias{dense_layer}
\title{Dense layer for final embedding}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
dense_layer
}
\description{
Dense layer for final embedding
}
\keyword{datasets}
