% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/speaker_encoder.R
\name{dense_layer}
\alias{dense_layer}
\title{Dense layer for final embedding}
\arguments{
\item{in_channels}{Input channels}

\item{out_channels}{Output channels}
}
\value{
nn_module
}
\description{
Dense layer for final embedding
}
