% tinyrox says don't edit this manually, but it can't stop you!
\name{download_chatterbox_models}
\alias{download_chatterbox_models}
\title{Download all chatterbox model files}
\usage{
download_chatterbox_models(force = FALSE)
}
\arguments{
\item{force}{Re-download all files}
}
\value{
Named list of local file paths
}
\description{
Download all chatterbox model files
}
