% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/download.R
\name{download_chatterbox_models}
\alias{download_chatterbox_models}
\title{Download all chatterbox model files}
\usage{
download_chatterbox_models(cache_dir = NULL, force = FALSE)
}
\arguments{
\item{cache_dir}{Cache directory}

\item{force}{Re-download all files}
}
\value{
Named list of local file paths
}
\description{
Download all chatterbox model files
}
