% tinyrox says don't edit this manually, but it can't stop you!
\name{drop_invalid_tokens}
\alias{drop_invalid_tokens}
\title{Drop invalid speech tokens}
\usage{
drop_invalid_tokens(tokens)
}
\arguments{
\item{tokens}{Token tensor}
}
\value{
Filtered tokens
}
\description{
Drop invalid speech tokens
}
