% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tokenizer.R
\name{drop_invalid_tokens}
\alias{drop_invalid_tokens}
\title{Drop invalid speech tokens}
\usage{
drop_invalid_tokens(tokens)
}
\arguments{
\item{tokens}{Token tensor}
}
\value{
Filtered tokens
}
\description{
Drop invalid speech tokens
}
