% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{espnet_rel_positional_encoding}
\alias{espnet_rel_positional_encoding}
\title{Sinusoidal positional encoding (Espnet RelPositionalEncoding)}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
espnet_rel_positional_encoding
}
\description{
Creates sinusoidal positional embeddings for use with relative position
attention. Includes scaling by sqrt(d_model) and adds positional embedding
to the input.
}
\keyword{datasets}
