% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/conformer.R
\name{espnet_rel_positional_encoding}
\alias{espnet_rel_positional_encoding}
\title{Sinusoidal positional encoding (Espnet RelPositionalEncoding)}
\usage{
espnet_rel_positional_encoding(
  d_model = 512,
  dropout_rate = 0.1,
  max_len = 5000
)
}
\arguments{
\item{d_model}{Model dimension}

\item{max_len}{Maximum sequence length}
}
\value{
nn_module
}
\description{
Creates sinusoidal positional embeddings for use with relative position
attention. Includes scaling by sqrt(d_model) and adds positional embedding
to the input.
}
