% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{fcm_module}
\alias{fcm_module}
\title{Factorized Convolutional Module (FCM)}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
fcm_module
}
\description{
Factorized Convolutional Module (FCM)
}
\keyword{datasets}
