% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/speaker_encoder.R
\name{fcm_module}
\alias{fcm_module}
\title{Factorized Convolutional Module (FCM)}
\arguments{
\item{m_channels}{Number of channels}

\item{feat_dim}{Input feature dimension (mel bins)}
}
\value{
nn_module
}
\description{
Factorized Convolutional Module (FCM)
}
