% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{feed_forward}
\alias{feed_forward}
\title{Feed-forward network for transformer
Matches diffusers FeedForward: net = [GELU(proj), Dropout, Linear]}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
feed_forward
}
\description{
Feed-forward network for transformer
Matches diffusers FeedForward: net = [GELU(proj), Dropout, Linear]
}
\keyword{datasets}
