% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{fsmn_multi_head_attention}
\alias{fsmn_multi_head_attention}
\title{FSMN Multi-Head Attention}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
fsmn_multi_head_attention
}
\description{
Multi-head attention with Frequency-domain Self-attention Memory Network
}
\keyword{datasets}
