% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tokenizer.R
\name{fsq_codebook}
\alias{fsq_codebook}
\title{FSQ Codebook module}
\usage{
fsq_codebook(dim, level = 3L)
}
\arguments{
\item{dim}{Input dimension (n_audio_state)}

\item{level}{Quantization level (default 3)}
}
\value{
nn_module
}
\description{
FSQ Codebook module
}
