% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{fsq_vector_quantization}
\alias{fsq_vector_quantization}
\title{FSQ Vector Quantization wrapper}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
fsq_vector_quantization
}
\description{
FSQ Vector Quantization wrapper
}
\keyword{datasets}
