% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{gelu_with_proj}
\alias{gelu_with_proj}
\title{GELU activation with projection (matches diffusers GELU structure)}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
gelu_with_proj
}
\description{
GELU activation with projection (matches diffusers GELU structure)
}
\keyword{datasets}
