% tinyrox says don't edit this manually, but it can't stop you!
\name{get_conv_padding}
\alias{get_conv_padding}
\title{Get padding for convolution}
\usage{
get_conv_padding(kernel_size, dilation = 1)
}
\arguments{
\item{kernel_size}{Kernel size}

\item{dilation}{Dilation rate}
}
\value{
Padding size
}
\description{
Get padding for convolution
}
