% tinyrox says don't edit this manually, but it can't stop you!
\name{get_turbo_model_paths}
\alias{get_turbo_model_paths}
\title{Get Paths to Downloaded Turbo Model Files}
\usage{
get_turbo_model_paths()
}
\value{
Named list of local file paths
}
\description{
Get Paths to Downloaded Turbo Model Files
}
