% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/hifigan.R
\name{hift_generator}
\alias{hift_generator}
\title{HiFTNet Generator}
\arguments{
\item{in_channels}{Input mel channels}

\item{base_channels}{Base channel count}

\item{nb_harmonics}{Number of harmonics for source filter}

\item{sampling_rate}{Output sample rate}

\item{nsf_alpha}{NSF sine amplitude}

\item{nsf_sigma}{NSF noise std}

\item{nsf_voiced_threshold}{F0 voiced threshold}

\item{upsample_rates}{Upsampling rates}

\item{upsample_kernel_sizes}{Upsampling kernel sizes}

\item{istft_n_fft}{ISTFT FFT size}

\item{istft_hop_len}{ISTFT hop length}

\item{resblock_kernel_sizes}{ResBlock kernel sizes}

\item{resblock_dilation_sizes}{ResBlock dilations}

\item{source_resblock_kernel_sizes}{Source resblock kernels}

\item{source_resblock_dilation_sizes}{Source resblock dilations}

\item{lrelu_slope}{LeakyReLU slope}

\item{audio_limit}{Output clipping limit}
}
\value{
nn_module
}
\description{
Neural Source Filter + ISTFTNet Reference:
https://arxiv.org/abs/2309.09493
}
