% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{learned_position_embeddings}
\alias{learned_position_embeddings}
\title{Learned position embeddings module}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
learned_position_embeddings
}
\description{
Learned position embeddings module
}
\keyword{datasets}
