% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{linear_no_subsampling}
\alias{linear_no_subsampling}
\title{Linear No Subsampling layer}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
linear_no_subsampling
}
\description{
Projects input to model dimension with layer norm and positional encoding.
}
\keyword{datasets}
