% tinyrox says don't edit this manually, but it can't stop you!
\name{llama_attention}
\alias{llama_attention}
\title{Llama attention module}
\usage{
llama_attention(config, layer_idx)
}
\arguments{
\item{config}{Model configuration}

\item{layer_idx}{Layer index}
}
\value{
nn_module
}
\description{
Llama attention module
}
