% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_attention}
\alias{llama_attention}
\title{Llama attention module}
\usage{
llama_attention(config, layer_idx)
}
\arguments{
\item{config}{Model configuration}

\item{layer_idx}{Layer index}
}
\value{
nn_module
}
\description{
Llama attention module
}
