% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{llama_attention}
\alias{llama_attention}
\title{Llama attention module}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
llama_attention
}
\description{
Llama attention module
}
\keyword{datasets}
