% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{llama_decoder_layer}
\alias{llama_decoder_layer}
\title{Llama decoder layer}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
llama_decoder_layer
}
\description{
Llama decoder layer
}
\keyword{datasets}
