% tinyrox says don't edit this manually, but it can't stop you!
\name{llama_decoder_layer}
\alias{llama_decoder_layer}
\title{Llama decoder layer}
\usage{
llama_decoder_layer(config, layer_idx)
}
\arguments{
\item{config}{Model configuration}

\item{layer_idx}{Layer index}
}
\value{
nn_module
}
\description{
Llama decoder layer
}
