% tinyrox says don't edit this manually, but it can't stop you!
\name{llama_model}
\alias{llama_model}
\title{Llama model (decoder only)}
\usage{
llama_model(config = NULL)
}
\arguments{
\item{config}{Model configuration (default: 520M)}
}
\value{
nn_module
}
\description{
Llama model (decoder only)
}
