% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{llama_model}
\alias{llama_model}
\title{Llama model (decoder only)}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
llama_model
}
\description{
Llama model (decoder only)
}
\keyword{datasets}
