% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_rms_norm}
\alias{llama_rms_norm}
\title{RMS Normalization module}
\usage{
llama_rms_norm(hidden_size, eps = 1e-05)
}
\arguments{
\item{hidden_size}{Dimension to normalize}

\item{eps}{Epsilon for numerical stability}
}
\value{
nn_module
}
\description{
RMS Normalization module
}
