% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{load_bpe_tokenizer}
\alias{load_bpe_tokenizer}
\title{Load tokenizer from JSON file}
\usage{
load_bpe_tokenizer(vocab_path)
}
\arguments{
\item{vocab_path}{Path to tokenizer.json}
}
\value{
Tokenizer object (list)
}
\description{
Load tokenizer from JSON file
}
