% tinyrox says don't edit this manually, but it can't stop you!
\name{load_chatterbox}
\alias{load_chatterbox}
\title{Load Chatterbox model weights}
\usage{
load_chatterbox(model, force = FALSE)
}
\arguments{
\item{model}{Chatterbox model object}

\item{force}{Force re-download of model files}
}
\value{
Chatterbox model with loaded weights
}
\description{
Downloads and loads pretrained weights for all model components.
}
