% tinyrox says don't edit this manually, but it can't stop you!
\name{load_s3gen_weights}
\alias{load_s3gen_weights}
\title{Load S3Gen weights}
\usage{
load_s3gen_weights(model, state_dict)
}
\arguments{
\item{model}{S3Gen model}

\item{state_dict}{State dictionary from safetensors}
}
\value{
Model with loaded weights
}
\description{
Load S3Gen weights
}
