% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tokenizer.R
\name{load_s3tokenizer_weights}
\alias{load_s3tokenizer_weights}
\title{Load S3Tokenizer weights}
\usage{
load_s3tokenizer_weights(model, state_dict)
}
\arguments{
\item{model}{S3Tokenizer model}

\item{state_dict}{Named list of tensors}
}
\value{
Model with loaded weights
}
\description{
Load S3Tokenizer weights
}
