% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t3.R
\name{load_t3_weights}
\alias{load_t3_weights}
\title{Load T3 weights from safetensors}
\usage{
load_t3_weights(model, state_dict)
}
\arguments{
\item{model}{T3 model}

\item{state_dict}{Named list of tensors}
}
\value{
Model with loaded weights
}
\description{
Load T3 weights from safetensors
}
