% tinyrox says don't edit this manually, but it can't stop you!
\name{load_voice_encoder_weights}
\alias{load_voice_encoder_weights}
\title{Load voice encoder weights from safetensors}
\usage{
load_voice_encoder_weights(model, state_dict)
}
\arguments{
\item{model}{Voice encoder model}

\item{state_dict}{Named list of tensors}
}
\value{
Model with loaded weights
}
\description{
Load voice encoder weights from safetensors
}
