% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tokenizer.R
\name{mask_to_bias}
\alias{mask_to_bias}
\title{Convert mask to attention bias}
\usage{
mask_to_bias(mask, dtype)
}
\arguments{
\item{mask}{Boolean mask}

\item{dtype}{Target dtype}
}
\value{
Attention bias tensor
}
\description{
Convert mask to attention bias
}
