% tinyrox says don't edit this manually, but it can't stop you!
\name{models_available}
\alias{models_available}
\title{Check if Models are Downloaded}
\usage{
models_available()
}
\value{
TRUE if all model files exist locally
}
\description{
Check if Models are Downloaded
}
\examples{
models_available()
}
