% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{models_available}
\alias{models_available}
\title{Check if models are downloaded}
\usage{
models_available(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Cache directory}
}
\value{
Logical indicating if all models are present
}
\description{
Check if models are downloaded
}
