% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/s3tokenizer.R
\name{pad_audio_for_tokenizer}
\alias{pad_audio_for_tokenizer}
\title{Pad audio to multiple of token rate}
\usage{
pad_audio_for_tokenizer(wav, sr)
}
\arguments{
\item{wav}{Audio samples}

\item{sr}{Sample rate}
}
\value{
Padded audio
}
\description{
Pad audio to multiple of token rate
}
