% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{perceiver_resampler}
\alias{perceiver_resampler}
\title{Perceiver resampler for conditioning compression}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
perceiver_resampler
}
\description{
Perceiver resampler for conditioning compression
}
\keyword{datasets}
