% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{positionwise_feedforward}
\alias{positionwise_feedforward}
\title{Positionwise Feed Forward}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
positionwise_feedforward
}
\description{
Two-layer feed-forward network with SiLU activation.
}
\keyword{datasets}
