% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{pre_lookahead_layer}
\alias{pre_lookahead_layer}
\title{Pre-Lookahead Layer}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
pre_lookahead_layer
}
\description{
Two causal convolutions with residual connection for look-ahead.
}
\keyword{datasets}
