% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tokenizer.R
\name{precompute_freqs_cis}
\alias{precompute_freqs_cis}
\title{Precompute rotary position embedding frequencies}
\usage{
precompute_freqs_cis(dim, end, theta = 10000)
}
\arguments{
\item{dim}{Dimension (head_dim)}

\item{end}{Maximum sequence length}

\item{theta}{Base frequency}
}
\value{
Complex frequency tensor
}
\description{
Precompute rotary position embedding frequencies
}
