% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.R
\name{print.chatterbox}
\alias{print.chatterbox}
\title{Print method for chatterbox}
\usage{
\method{print}{chatterbox}(x, ...)
}
\arguments{
\item{x}{Chatterbox model}

\item{...}{Ignored}
}
\description{
Print method for chatterbox
}
