% tinyrox says don't edit this manually, but it can't stop you!
\name{print.voice_embedding}
\alias{print.voice_embedding}
\title{Print method for voice_embedding}
\usage{
\method{print}{voice_embedding}(x, ...)
}
\arguments{
\item{x}{Voice embedding}

\item{...}{Ignored}
}
\description{
Print method for voice_embedding
}
