% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.R
\name{print.voice_embedding}
\alias{print.voice_embedding}
\title{Print method for voice_embedding}
\usage{
\method{print}{voice_embedding}(x, ...)
}
\arguments{
\item{x}{Voice embedding}

\item{...}{Ignored}
}
\description{
Print method for voice_embedding
}
