% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_utils.R
\name{read_audio}
\alias{read_audio}
\title{Read audio file}
\usage{
read_audio(path)
}
\arguments{
\item{path}{Path to audio file (WAV format)}
}
\value{
List with samples (numeric vector normalized to [-1, 1]) and sr (sample rate)
}
\description{
Read audio file
}
