% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetensors.R
\name{read_safetensors}
\alias{read_safetensors}
\title{Read safetensors file}
\usage{
read_safetensors(path, device = "cpu")
}
\arguments{
\item{path}{Path to .safetensors file}

\item{device}{Device to load tensors to ("cpu", "cuda", etc.)}
}
\value{
Named list of torch tensors
}
\description{
Read safetensors file
}
