% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/hifigan.R
\name{reflection_pad1d}
\alias{reflection_pad1d}
\title{Reflection padding for 1D (nn_reflection_pad1d equivalent)}
\usage{
reflection_pad1d(padding)
}
\arguments{
\item{padding}{Integer vector c(left, right) for padding}
}
\value{
nn_module
}
\description{
Reflection padding for 1D (nn_reflection_pad1d equivalent)
}
