% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{reflection_pad1d}
\alias{reflection_pad1d}
\title{Reflection padding for 1D (nn_reflection_pad1d equivalent)}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
reflection_pad1d
}
\description{
Reflection padding for 1D (nn_reflection_pad1d equivalent)
}
\keyword{datasets}
