% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{rel_position_attention}
\alias{rel_position_attention}
\title{Relative Position Multi-Headed Attention}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
rel_position_attention
}
\description{
Multi-head attention with relative positional encodings.
}
\keyword{datasets}
