% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_utils.R
\name{resample_audio}
\alias{resample_audio}
\title{Resample audio}
\usage{
resample_audio(samples, from_sr, to_sr)
}
\arguments{
\item{samples}{Numeric vector of audio samples}

\item{from_sr}{Source sample rate}

\item{to_sr}{Target sample rate}
}
\value{
Resampled audio samples
}
\description{
Resample audio
}
