% tinyrox says don't edit this manually, but it can't stop you!
\name{resample_audio}
\alias{resample_audio}
\title{Resample audio}
\usage{
resample_audio(samples, from_sr, to_sr)
}
\arguments{
\item{samples}{Numeric vector of audio samples}

\item{from_sr}{Source sample rate}

\item{to_sr}{Target sample rate}
}
\value{
Resampled audio samples
}
\description{
Resample audio
}
