% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{s3_audio_encoder}
\alias{s3_audio_encoder}
\title{S3 Audio Encoder V2}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
s3_audio_encoder
}
\description{
S3 Audio Encoder V2
}
\keyword{datasets}
