% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/s3tokenizer.R
\name{s3_multi_head_attention}
\alias{s3_multi_head_attention}
\title{Multi-Head Attention base module}
\arguments{
\item{n_state}{Hidden dimension}

\item{n_head}{Number of heads}
}
\value{
nn_module
}
\description{
Multi-Head Attention base module
}
