% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{s3_multi_head_attention}
\alias{s3_multi_head_attention}
\title{Multi-Head Attention base module}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
s3_multi_head_attention
}
\description{
Multi-Head Attention base module
}
\keyword{datasets}
