% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{s3_residual_attention_block}
\alias{s3_residual_attention_block}
\title{Residual attention block}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
s3_residual_attention_block
}
\description{
Residual attention block
}
\keyword{datasets}
