% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tokenizer.R
\name{s3_residual_attention_block}
\alias{s3_residual_attention_block}
\title{Residual attention block}
\usage{
s3_residual_attention_block(n_state, n_head, kernel_size = 31L)
}
\arguments{
\item{n_state}{Hidden dimension}

\item{n_head}{Number of heads}

\item{kernel_size}{FSMN kernel size}
}
\value{
nn_module
}
\description{
Residual attention block
}
