% tinyrox says don't edit this manually, but it can't stop you!
\name{s3_tokenizer}
\alias{s3_tokenizer}
\title{S3Tokenizer V2 module}
\usage{
s3_tokenizer(config = NULL)
}
\arguments{
\item{config}{Configuration list (default from s3_tokenizer_config())}
}
\value{
nn_module
}
\description{
S3Tokenizer V2 module
}
