% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{s3_tokenizer}
\alias{s3_tokenizer}
\title{S3Tokenizer V2 module}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
s3_tokenizer
}
\description{
S3Tokenizer V2 module
}
\keyword{datasets}
