% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/hifigan.R
\name{sine_gen}
\alias{sine_gen}
\title{Sine Generator}
\arguments{
\item{sample_rate}{Sampling rate in Hz}

\item{harmonic_num}{Number of harmonics}

\item{sine_amp}{Sine amplitude}

\item{noise_std}{Noise standard deviation}

\item{voiced_threshold}{F0 threshold for voiced/unvoiced}
}
\value{
nn_module
}
\description{
Generates sine waveforms from F0 for source-filter synthesis
}
