% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{sinusoidal_pos_emb}
\alias{sinusoidal_pos_emb}
\title{Sinusoidal positional embedding for timesteps}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
sinusoidal_pos_emb
}
\description{
Sinusoidal positional embedding for timesteps
}
\keyword{datasets}
