% tinyrox says don't edit this manually, but it can't stop you!
\name{sinusoidal_pos_emb}
\alias{sinusoidal_pos_emb}
\title{Sinusoidal positional embedding for timesteps}
\usage{
sinusoidal_pos_emb(dim = 320L)
}
\arguments{
\item{dim}{Output dimension}
}
\value{
nn_module
}
\description{
Sinusoidal positional embedding for timesteps
}
