% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/hifigan.R
\name{snake_activation}
\alias{snake_activation}
\title{Snake activation function}
\usage{
snake_activation(in_features, alpha_trainable = TRUE, alpha_logscale = FALSE)
}
\arguments{
\item{in_features}{Number of input channels}

\item{alpha_trainable}{Whether alpha is trainable}

\item{alpha_logscale}{Whether to use log scale for alpha}
}
\value{
nn_module
}
\description{
Sine-based periodic activation: x + 1/a * sin^2(ax)
Reference: https://arxiv.org/abs/2006.08195
}
