% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hifigan.R
\name{source_module_hn_nsf}
\alias{source_module_hn_nsf}
\title{Source Module for Neural Source Filter}
\usage{
source_module_hn_nsf(
  sample_rate,
  upsample_scale,
  harmonic_num = 0,
  sine_amp = 0.1,
  add_noise_std = 0.003,
  voiced_threshold = 0
)
}
\arguments{
\item{sample_rate}{Sampling rate}

\item{upsample_scale}{Upsampling factor}

\item{harmonic_num}{Number of harmonics}

\item{sine_amp}{Sine amplitude}

\item{add_noise_std}{Noise std}

\item{voiced_threshold}{Voiced threshold}
}
\value{
nn_module
}
\description{
Source Module for Neural Source Filter
}
