% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speaker_encoder.R
\name{statistics_pooling}
\alias{statistics_pooling}
\title{Statistics pooling}
\usage{
statistics_pooling(x)
}
\arguments{
\item{x}{Input tensor (batch, channels, time)}
}
\value{
Statistics tensor (batch, channels * 2)
}
\description{
Statistics pooling
}
