% tinyrox says don't edit this manually, but it can't stop you!
\name{t3_cond_to_device}
\alias{t3_cond_to_device}
\title{Move T3 conditioning to device}
\usage{
t3_cond_to_device(cond, device)
}
\arguments{
\item{cond}{T3 conditioning object}

\item{device}{Target device}
}
\value{
T3 conditioning on device
}
\description{
Move T3 conditioning to device
}
