% tinyrox says don't edit this manually, but it can't stop you!
\name{timestep_embedding}
\alias{timestep_embedding}
\title{Timestep embedding MLP}
\usage{
timestep_embedding(in_channels = 320L, time_embed_dim = 1024L)
}
\arguments{
\item{in_channels}{Input channels}

\item{time_embed_dim}{Output dimension}
}
\value{
nn_module
}
\description{
Timestep embedding MLP
}
