% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{traceable_decoder_layer}
\alias{traceable_decoder_layer}
\title{Traceable decoder layer with pre-allocated KV cache}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
traceable_decoder_layer
}
\description{
Traceable decoder layer with pre-allocated KV cache
}
\keyword{datasets}
