% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{traceable_kv_projector}
\alias{traceable_kv_projector}
\title{Traceable K/V projection module}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
traceable_kv_projector
}
\description{
Computes K and V projections with RoPE for a single layer.
Returns concatenated K and V for easy unpacking.
}
\keyword{datasets}
