% tinyrox says don't edit this manually, but it can't stop you!
\name{traceable_transformer_first}
\alias{traceable_transformer_first}
\title{Traceable transformer for first token (no cache)}
\usage{
traceable_transformer_first(tfmr)
}
\arguments{
\item{tfmr}{Original llama_model}
}
\value{
nn_module
}
\description{
Traceable transformer for first token (no cache)
}
