% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/llama_traced.R
\name{traceable_transformer_first}
\alias{traceable_transformer_first}
\title{Traceable transformer for first token (no cache)}
\usage{
traceable_transformer_first(tfmr)
}
\arguments{
\item{tfmr}{Original llama_model}
}
\value{
nn_module
}
\description{
Traceable transformer for first token (no cache)
}
