% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speaker_encoder.R
\name{transit_layer}
\alias{transit_layer}
\title{Transit layer (channel reduction)}
\usage{
transit_layer(in_channels, out_channels, bias = FALSE)
}
\arguments{
\item{in_channels}{Input channels}

\item{out_channels}{Output channels}

\item{bias}{Whether to use bias (default FALSE)}
}
\value{
nn_module
}
\description{
Transit layer (channel reduction)
}
