% tinyrox says don't edit this manually, but it can't stop you!
\name{turbo_models_available}
\alias{turbo_models_available}
\title{Check if Turbo Models are Downloaded}
\usage{
turbo_models_available()
}
\value{
TRUE if all turbo model files exist locally
}
\description{
Check if Turbo Models are Downloaded
}
\examples{
turbo_models_available()
}
