% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{upsample_1d}
\alias{upsample_1d}
\title{Upsample 1D}
\format{
An object of class \code{nn_module_generator} of length 1.
}
\usage{
upsample_1d
}
\description{
2x upsampling using interpolation + convolution.
}
\keyword{datasets}
