% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/s3gen.R
\name{upsample_conformer_encoder}
\alias{upsample_conformer_encoder}
\title{Upsample Conformer Encoder}
\arguments{
\item{input_size}{Input dimension}

\item{output_size}{Output dimension}

\item{num_blocks}{Number of conformer blocks}
}
\value{
nn_module
}
\description{
Upsample Conformer Encoder
}
