% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voice_encoder.R
\name{voice_encoder}
\alias{voice_encoder}
\title{Voice encoder module}
\usage{
voice_encoder(config = NULL)
}
\arguments{
\item{config}{Voice encoder configuration}
}
\value{
nn_module
}
\description{
Voice encoder module
}
