% tinyrox says don't edit this manually, but it can't stop you!
\name{write_audio}
\alias{write_audio}
\title{Write audio file}
\usage{
write_audio(samples, sr, path)
}
\arguments{
\item{samples}{Numeric vector of audio samples (normalized to \[-1, 1\])}

\item{sr}{Sample rate}

\item{path}{Output path (WAV format)}
}
\value{
The output \code{path}, invisibly. Called for the side effect
  of writing a WAV file.
}
\description{
Write audio file
}
