% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/audio_utils.R
\name{write_audio}
\alias{write_audio}
\title{Write audio file}
\usage{
write_audio(samples, sr, path)
}
\arguments{
\item{samples}{Numeric vector of audio samples (normalized to \[-1, 1\])}

\item{sr}{Sample rate}

\item{path}{Output path (WAV format)}
}
\description{
Write audio file
}
